Split session transport tests into smaller modules
